feat: não permite excluir categoria vinculada a um produto

+++ backend/src/useCases/categories/delete-category-usecase.ts
@@ -1,4 +1,6 @@
+import { ForeignKeyConstraintError } from "../../errors/foreign-key-constraint-error";
 import { CategoryRepository } from "../../repositories/category-repository";
+import { ProductRepository } from "../../repositories/product-repository";
 import { Category } from "../../types/category";
 
 interface DeleteCategoryUseCaseRequest {
@@ -10,9 +12,18 @@ interface DeleteCategoryUseCaseResponse {
 }
 
 export class DeleteCategoryUseCase {
-    constructor(private categoryRepository: CategoryRepository) { }
+    constructor(
+        private categoryRepository: CategoryRepository
+    ) { }
 
     async execute({ id }: DeleteCategoryUseCaseRequest): Promise<DeleteCategoryUseCaseResponse> {
+
+        const categoryInUse = await this.categoryRepository.listOneProduct(id)
+
+        if (categoryInUse) {
+            throw new ForeignKeyConstraintError()
+        }
+
         const category = await this.categoryRepository.delete(id)
 
         return {
